pyproject.toml
[project]
name = "oenskeskyen-batch-upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
]

[tool.pyright]
include = ["src", "app.py", "main.py", "start_api.py"]
pythonPath = "."

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
